requirements/base.txt
pyjwt[crypto]
itsdangerous
aiosmtplib
aiohttp==3.10.10
alembic
annotated-types
apscheduler==3.10.4
argon2-cffi
asgiref
asyncpg
babel==2.16.0
bcrypt==4.0.1
ccxt==4.4.23
celery
cloudinary
dnspython
fastapi[all]==0.115.4
fastapi-pagination==0.12.31
hiredis
humanize
loguru
passlib
phonenumbers==8.13.47
pillow
pycountry==24.6.1
pydantic
pydantic-settings
pydantic-extra-types==2.9.0
pysdk-sui==5.1
pysui==0.71.0
python-dateutil
python-jose
python-magic
python-multipart
python-slugify
python-telegram-bot==21.6
redis
resend
requests
sqlalchemy[asyncio]
sqlmodel==0.0.22
sui-python-sdk==0.1.5 # Generate sui wallets and sign transactions with this
typing-extensions
urllib3
websockets